+++ packages/aiinfo/node/components/igniteagent/index.ts
@@ -1,8 +1,8 @@
-import {
-  type AiInfoFeatureProps,
+import  {
   type AiInformationProps,
   type DataPermissionLevelsProps,
   type NutritionFactsProps,
+  type AiInfoFeatureProps,
 } from "../../types";
 const FEATURE_NAME = "Agent";
 const UID = "igniteagent";
@@ -43,7 +43,7 @@ const NUTRITION_FACTS_DATA: NutritionFactsProps["data"] = [
       {
         description: "The foundational AI on which further training and customizations are built.",
         segmentTitle: "Base Model",
-        value: "Haiku 3, Sonnet 3.7",
+        value: "Haiku 3, Haiku 3.5, Haiku 4.5, Sonnet 3.7",
         valueDescription:
           "Anthropic Claude models are provided via Instructure's in-house AI Platform.",
       },
@@ -150,7 +150,7 @@ const aiInformation: AiInformationProps = {
       description:
         "We utilize off-the-shelf AI models and customer data as input to provide AI-powered features. No data is used for training this model.",
       featureName: "Agent",
-      modelName: "Haiku 3, Sonnet 3.7",
+      modelName: "Haiku 3, Haiku 3.5, Haiku 4.5, Sonnet 3.7",
       modelNameText: "Base Model",
       nutritionFactsModalTriggerText: "AI Nutrition Facts",
       permissionLevel: "LEVEL 2",